requirements.txt
gunicorn
psycopg2-binary
django
djangorestframework